samples/bbq-beach-agent/pyproject.toml
[project]
name = "bbq-beach-agent"
version = "0.1.0"
description = "BBQのできるビーチを検索するAIエージェント"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.2",
    "click>=8.2.0",
    "geopy>=2.4.1",
    "google-adk>=1.0.0",
    "gradio>=5.30.0",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.1.1",
    "langgraph>=0.4.5",
    "mcp>=1.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bbq_beach_agent"]

[project.scripts]
bbq-beach-agent = "bbq_beach_agent.__main__:cli_main"
